requirements.txt
hydra-core
pytest
torch
